+++ src/sider/oversikt/filter/ClearFiltersButton.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from 'react';
 import styled from 'styled-components';
 import themes from '@/styles/themes';
-import SlettIkon from '@/components/common/SlettIkon';
+import SlettIkon from '@/sider/oversikt/filter/SlettIkon';
 import { useFilters } from '@/context/filters/FilterContext';
 import { ActionType } from '@/context/filters/filterContextActions';
 
@@ -20,12 +20,15 @@ const ClearFiltersWrapper = styled.div`
   padding: 0.5rem 0.25rem;
   -webkit-user-select: none;
   user-select: none;
+
   :hover {
     cursor: pointer;
   }
+
   :active {
     border: 2px solid ${themes.color.navBla};
   }
+
   > svg {
     margin-right: 0.25em;
   }
