+++ src/sider/oversikt/filter/SlettIkon.tsx
