+++ src/sider/oversikt/sokeresultat/oversikttable/PersonRadVirksomhetColumn.tsx
